Numeric.Additive.Class

Additive Semigroups

class Additive r where

(a + b) + c = a + (b + c)
sinnum 1 a = a
sinnum (2 * n) a = sinnum n a + sinnum n a
sinnum (2 * n + 1) a = sinnum n a + sinnum n a + a

Minimal complete definition

(+)

Methods

(+) :: r -> r -> r infixl 6

sinnum1p :: Natural -> r -> r

sinnum1p n r = sinnum (1 + n) r

sumWith1 :: Foldable1 f => (a -> r) -> f a -> r

sum1 :: (Foldable1 f, Additive r) => f r -> r

Additive Abelian semigroups

class Additive r => Abelian r

an additive abelian semigroup

a + b = b + a

Additive Monoids

class Additive r => Idempotent r

An additive semigroup with idempotent addition.

a + a = a

sinnum1pIdempotent :: Natural -> r -> r

Partitionable semigroups

class Additive m => Partitionable m where

Minimal complete definition

partitionWith

Methods

partitionWith :: (m -> m -> r) -> m -> NonEmpty r

partitionWith f c returns a list containing f a b for each a b such that a + b = c,