Numeric.Algebra

Additive

additive semigroups

class Additive r where

(a + b) + c = a + (b + c)
sinnum 1 a = a
sinnum (2 * n) a = sinnum n a + sinnum n a
sinnum (2 * n + 1) a = sinnum n a + sinnum n a + a

Minimal complete definition

(+)

Methods

(+) :: r -> r -> r infixl 6

sinnum1p :: Natural -> r -> r

sinnum1p n r = sinnum (1 + n) r

sumWith1 :: Foldable1 f => (a -> r) -> f a -> r

sum1 :: (Foldable1 f, Additive r) => f r -> r

additive Abelian semigroups

class Additive r => Abelian r

an additive abelian semigroup

a + b = b + a

additive idempotent semigroups

class Additive r => Idempotent r

An additive semigroup with idempotent addition.

a + a = a

sinnum1pIdempotent :: Natural -> r -> r

sinnumIdempotent :: (Integral n, Idempotent r, Monoidal r) => n -> r -> r

partitionable additive semigroups

class Additive m => Partitionable m where

Minimal complete definition

partitionWith

Methods

partitionWith :: (m -> m -> r) -> m -> NonEmpty r

partitionWith f c returns a list containing f a b for each a b such that a + b = c,

additive monoids

class (LeftModuleNatural m, RightModuleNatural m) => Monoidal m where

An additive monoid

zero + a = a = a + zero

Minimal complete definition

zero

Methods

zero :: m

sinnum :: Natural -> m -> m

sumWith :: Foldable f => (a -> m) -> f a -> m

sum :: (Foldable f, Monoidal m) => f m -> m

additive groups

class (LeftModuleInteger r, RightModuleInteger r, Monoidal r) => Group r where

Methods

(-) :: r -> r -> r infixl 6

negate :: r -> r

subtract :: r -> r -> r

times :: Integral n => n -> r -> r infixl 7

Multiplicative

multiplicative semigroups

class Multiplicative r where

A multiplicative semigroup

Minimal complete definition

(*)

Methods

(*) :: r -> r -> r infixl 7

pow1p :: r -> Natural -> r infixr 8

productWith1 :: Foldable1 f => (a -> r) -> f a -> r

product1 :: (Foldable1 f, Multiplicative r) => f r -> r

commutative multiplicative semigroups

class Multiplicative r => Commutative r

A commutative multiplicative semigroup

multiplicative monoids

class Multiplicative r => Unital r where

Minimal complete definition

one

Methods

one :: r

pow :: r -> Natural -> r infixr 8

productWith :: Foldable f => (a -> r) -> f a -> r

product :: (Foldable f, Unital r) => f r -> r

idempotent multiplicative semigroups

class Multiplicative r => Band r

An multiplicative semigroup with idempotent multiplication.

a * a = a

pow1pBand :: r -> Natural -> r

powBand :: Unital r => r -> Natural -> r

multiplicative groups

class Unital r => Division r where

Methods

recip :: r -> r

(/) :: r -> r -> r infixl 7

(\\) :: r -> r -> r infixl 7

(^) :: Integral n => r -> n -> r infixr 8

factorable multiplicative semigroups

class Multiplicative m => Factorable m where

`factorWith f c` returns a non-empty list containing `f a b` for all `a, b` such that `a * b = c`.

Results of factorWith f 0 are undefined and may result in either an error or an infinite list.

Minimal complete definition

factorWith

Methods

factorWith :: (m -> m -> r) -> m -> NonEmpty r

involutive multiplicative semigroups

class Multiplicative r => InvolutiveMultiplication r where

An semigroup with involution

adjoint a * adjoint b = adjoint (b * a)

Minimal complete definition

adjoint

Methods

adjoint :: r -> r

class (Commutative r, InvolutiveMultiplication r) => TriviallyInvolutive r

adjoint = id

Ring-Structures

Semirings

class (Additive r, Abelian r, Multiplicative r) => Semiring r

A pair of an additive abelian semigroup, and a multiplicative semigroup, with the distributive laws:

a(b + c) = ab + ac -- left distribution (we are a LeftNearSemiring)
(a + b)c = ac + bc -- right distribution (we are a [Right]NearSemiring)

Common notation includes the laws for additive and multiplicative identity in semiring.

If you want that, look at Rig instead.

Ideally we'd use the cyclic definition:

class (LeftModule r r, RightModule r r, Additive r, Abelian r, Multiplicative r) => Semiring r

to enforce that every semiring r is an r-module over itself, but Haskell doesn't like that.

class (Semiring r, InvolutiveMultiplication r) => InvolutiveSemiring r

adjoint (x + y) = adjoint x + adjoint y

class (Semiring r, Idempotent r) => Dioid r

Rngs

class (Group r, Semiring r) => Rng r

A Ring without an identity.

Rigs

class (Semiring r, Unital r, Monoidal r) => Rig r where

A Ring without (n)egation

Methods

fromNatural :: Natural -> r

Rings

class (Rig r, Rng r) => Ring r where

Methods

fromInteger :: Integer -> r

Division Rings

class Ring r => LocalRing r

class (Division r, Ring r) => DivisionRing r

class (Euclidean d, Division d) => Field d

Modules

class (Semiring r, Additive m) => LeftModule r m where

Minimal complete definition

(.*)

Methods

(.*) :: r -> m -> m infixl 7

class (Semiring r, Additive m) => RightModule r m where

Minimal complete definition

(*.)

Methods

(*.) :: m -> r -> m infixl 7

class (LeftModule r m, RightModule r m) => Module r m

Algebras

associative algebras over (non-commutative) semirings

class Semiring r => Algebra r a where

An associative algebra built with a free module over a semiring

Minimal complete definition

mult

Methods

mult :: (a -> a -> r) -> a -> r

class Semiring r => Coalgebra r c where

Minimal complete definition

comult

Methods

comult :: (c -> r) -> c -> c -> r

unital algebras

class Algebra r a => UnitalAlgebra r a where

An associative unital algebra over a semiring, built using a free module

Minimal complete definition

unit

Methods

unit :: r -> a -> r

class Coalgebra r c => CounitalCoalgebra r c where

Minimal complete definition

counit

Methods

counit :: (c -> r) -> r

class (UnitalAlgebra r a, CounitalCoalgebra r a) => Bialgebra r a

A bialgebra is both a unital algebra and counital coalgebra where the mult and unit are compatible in some sense with the comult and counit. That is to say that mult and unit are a coalgebra homomorphisms or (equivalently) that comult and counit are an algebra homomorphisms.

involutive algebras

class (InvolutiveSemiring r, Algebra r a) => InvolutiveAlgebra r a where

Minimal complete definition

inv

Methods

inv :: (a -> r) -> a -> r

class (InvolutiveSemiring r, Coalgebra r c) => InvolutiveCoalgebra r c where

Minimal complete definition

coinv

Methods

coinv :: (c -> r) -> c -> r

class (Bialgebra r h, InvolutiveAlgebra r h, InvolutiveCoalgebra r h) => InvolutiveBialgebra r h

class (CommutativeAlgebra r a, TriviallyInvolutive r, InvolutiveAlgebra r a) => TriviallyInvolutiveAlgebra r a

class (CocommutativeCoalgebra r a, TriviallyInvolutive r, InvolutiveCoalgebra r a) => TriviallyInvolutiveCoalgebra r a

class (InvolutiveBialgebra r h, TriviallyInvolutiveAlgebra r h, TriviallyInvolutiveCoalgebra r h) => TriviallyInvolutiveBialgebra r h

idempotent algebras

class Algebra r a => IdempotentAlgebra r a

class (Bialgebra r h, IdempotentAlgebra r h, IdempotentCoalgebra r h) => IdempotentBialgebra r h

commutative algebras

class Algebra r a => CommutativeAlgebra r a

class (Bialgebra r h, CommutativeAlgebra r h, CocommutativeCoalgebra r h) => CommutativeBialgebra r h

class Coalgebra r c => CocommutativeCoalgebra r c

division algebras

class UnitalAlgebra r a => DivisionAlgebra r a where

Minimal complete definition

recipriocal

Methods

recipriocal :: (a -> r) -> a -> r

Hopf alegebras

class Bialgebra r h => HopfAlgebra r h where

A HopfAlgebra algebra on a semiring, where the module is free.

When antipode . antipode = id and antipode is an antihomomorphism then we are an InvolutiveBialgebra with inv = antipode as well

Minimal complete definition

antipode

Methods

antipode :: (h -> r) -> h -> r

Ring Properties

Characteristic

class Rig r => Characteristic r where

Minimal complete definition

char

Methods

char :: proxy r -> Natural

charInt :: (Integral s, Bounded s) => proxy s -> Natural

charWord :: (Integral s, Bounded s) => proxy s -> Natural

Order

class Order a where

Methods

(<~) :: a -> a -> Bool

(<) :: a -> a -> Bool

(>~) :: a -> a -> Bool

(>) :: a -> a -> Bool

(~~) :: a -> a -> Bool

(/~) :: a -> a -> Bool

order :: a -> a -> MaybeOrdering

comparable :: a -> a -> Bool

class (AdditiveOrder r, Rig r) => OrderedRig r

class (Additive r, Order r) => AdditiveOrder r

z + x <= z + y = x <= y = x + z <= y + z

class Order a => LocallyFiniteOrder a

Minimal complete definition

range, rangeSize

class Monoidal r => DecidableZero r

Minimal complete definition

isZero

class Unital r => DecidableUnits r

Minimal complete definition

recipUnit

class Unital r => DecidableAssociates r

Minimal complete definition

isAssociate

Natural numbers

data Natural :: *

Type representing arbitrary-precision non-negative integers.

Operations whose result would be negative throw (Underflow :: ArithException).

Since: 4.8.0.0

Representable Additive

addRep :: (Applicative m, Additive r) => m r -> m r -> m r

`Additive.(+)` default definition

sinnum1pRep :: (Functor m, Additive r) => Natural -> m r -> m r

sinnum1p default definition

Representable Monoidal

zeroRep :: (Applicative m, Monoidal r) => m r

zero default definition

sinnumRep :: (Functor m, Monoidal r) => Natural -> m r -> m r

sinnum default definition

Representable Group

negateRep :: (Functor m, Group r) => m r -> m r

negate default definition

minusRep :: (Applicative m, Group r) => m r -> m r -> m r

`Group.(-)` default definition

subtractRep :: (Applicative m, Group r) => m r -> m r -> m r

subtract default definition

timesRep :: (Integral n, Functor m, Group r) => n -> m r -> m r

times default definition

Representable Multiplicative (via Algebra)

mulRep :: (Representable m, Algebra r (Rep m)) => m r -> m r -> m r

`Multiplicative.(*)` default definition

Representable Unital (via UnitalAlgebra)

oneRep :: (Representable m, Unital r, UnitalAlgebra r (Rep m)) => m r

one default definition

Representable Rig (via Algebra)

fromNaturalRep :: (UnitalAlgebra r (Rep m), Representable m, Rig r) => Natural -> m r

fromNatural default definition

Representable Ring (via Algebra)

fromIntegerRep :: (UnitalAlgebra r (Rep m), Representable m, Ring r) => Integer -> m r

fromInteger default definition

Norm

class Additive r => Quadrance r m where

Minimal complete definition

quadrance

Methods

quadrance :: m -> r

Covectors

newtype Covector r a

Linear functionals from elements of an (infinite) free module to a scalar

Constructors

Covector

Fields

($*) :: (a -> r) -> r

Covectors as linear functionals

counitM :: UnitalAlgebra r a => a -> Covector r ()

unitM :: CounitalCoalgebra r c => Covector r c

comultM :: Algebra r a => a -> Covector r (a, a)

multM :: Coalgebra r c => c -> c -> Covector r c

invM :: InvolutiveAlgebra r h => h -> Covector r h

coinvM :: InvolutiveCoalgebra r h => h -> Covector r h

antipodeM :: HopfAlgebra r h => h -> Covector r h

convolveM antipodeM return = convolveM return antipodeM = comultM >=> uncurry joinM

convolveM :: (Algebra r c, Coalgebra r a) => (c -> Covector r a) -> (c -> Covector r a) -> c -> Covector r a