Numeric.Algebra.Class

Multiplicative Semigroups

class Multiplicative r where

A multiplicative semigroup

Minimal complete definition

(*)

Methods

(*) :: r -> r -> r infixl 7

pow1p :: r -> Natural -> r infixr 8

productWith1 :: Foldable1 f => (a -> r) -> f a -> r

pow1pIntegral :: (Integral r, Integral n) => r -> n -> r

product1 :: (Foldable1 f, Multiplicative r) => f r -> r

Semirings

class (Additive r, Abelian r, Multiplicative r) => Semiring r

A pair of an additive abelian semigroup, and a multiplicative semigroup, with the distributive laws:

a(b + c) = ab + ac -- left distribution (we are a LeftNearSemiring)
(a + b)c = ac + bc -- right distribution (we are a [Right]NearSemiring)

Common notation includes the laws for additive and multiplicative identity in semiring.

If you want that, look at Rig instead.

Ideally we'd use the cyclic definition:

class (LeftModule r r, RightModule r r, Additive r, Abelian r, Multiplicative r) => Semiring r

to enforce that every semiring r is an r-module over itself, but Haskell doesn't like that.

Left and Right Modules

class (Semiring r, Additive m) => LeftModule r m where

Minimal complete definition

(.*)

Methods

(.*) :: r -> m -> m infixl 7

class (Semiring r, Additive m) => RightModule r m where

Minimal complete definition

(*.)

Methods

(*.) :: m -> r -> m infixl 7

class (LeftModule r m, RightModule r m) => Module r m

Additive Monoids

class (LeftModuleNatural m, RightModuleNatural m) => Monoidal m where

An additive monoid

zero + a = a = a + zero

Minimal complete definition

zero

Methods

zero :: m

sinnum :: Natural -> m -> m

sumWith :: Foldable f => (a -> m) -> f a -> m

sum :: (Foldable f, Monoidal m) => f m -> m

sinnumIdempotent :: (Integral n, Idempotent r, Monoidal r) => n -> r -> r

Associative algebras

class Semiring r => Algebra r a where

An associative algebra built with a free module over a semiring

Minimal complete definition

mult

Methods

mult :: (a -> a -> r) -> a -> r

Coassociative coalgebras

class Semiring r => Coalgebra r c where

Minimal complete definition

comult

Methods

comult :: (c -> r) -> c -> c -> r