Numeric.Algebra.Unital

Unital Multiplication (Multiplicative monoid)

class Multiplicative r => Unital r where

Minimal complete definition

one

Methods

one :: r

pow :: r -> Natural -> r infixr 8

productWith :: Foldable f => (a -> r) -> f a -> r

product :: (Foldable f, Unital r) => f r -> r

Unital Associative Algebra

class Algebra r a => UnitalAlgebra r a where

An associative unital algebra over a semiring, built using a free module

Minimal complete definition

unit

Methods

unit :: r -> a -> r

Unital Coassociative Coalgebra

class Coalgebra r c => CounitalCoalgebra r c where

Minimal complete definition

counit

Methods

counit :: (c -> r) -> r

Bialgebra

class (UnitalAlgebra r a, CounitalCoalgebra r a) => Bialgebra r a

A bialgebra is both a unital algebra and counital coalgebra where the mult and unit are compatible in some sense with the comult and counit. That is to say that mult and unit are a coalgebra homomorphisms or (equivalently) that comult and counit are an algebra homomorphisms.