Numeric.Decidable.Associates

Documentation

class Unital r => DecidableAssociates r where

Minimal complete definition

isAssociate

Methods

isAssociate :: r -> r -> Bool

b is an associate of a if there exists a unit u such that b = a*u

This relationship is symmetric because if u is a unit, u^-1 exists and is a unit, so

b*u^-1 = a*u*u^-1 = a

isAssociateIntegral :: (Eq n, Num n) => n -> n -> Bool

isAssociateWhole :: Eq n => n -> n -> Bool