Numeric.Ring.Rng

Documentation

data RngRing r

The free Ring given a Rng obtained by adjoining Z, such that

RngRing r = n*1 + r

This ring is commonly denoted r^.

Constructors

RngRing !Integer r

rngRingHom :: r -> RngRing r

The rng homomorphism from r to RngRing r

liftRngHom :: Ring s => (r -> s) -> RngRing r -> s

given a rng homomorphism from a rng r into a ring s, liftRngHom yields a ring homomorphism from the ring `r^` into s.